delete syntax package
